Fix; finish rebase and implement Payment

+++ bftengine/include/bftengine/ClientMsgs.hpp
@@ -34,12 +34,8 @@ struct ClientBatchRequestMsgHeader {
 struct ClientRequestMsgHeader {
   uint16_t msgType;  // always == REQUEST_MSG_TYPE
   uint32_t spanContextSize = 0u;
-  uint16_t idOfClientProxy;
-  // bit 0 == isReadOnly, 
-  // bit 1 = preProcess, 
-  // bit 2 = quickpay, 
-  // bits 3-7 are reserved
-  uint8_t flags;  
+  uint16_t idOfClientProxy;  // TODO - rename - now used mostly as id of external client
+  uint64_t flags;            // bit 0 == isReadOnly, bit 1 = preProcess, bits 2-7 are reserved
   uint64_t reqSeqNum;
   uint32_t requestLength;
   uint64_t timeoutMilli;
